v27

+++ app/src/main/java/com/example/notes_app/modul/room_database/data_classes/DiaryHashtagJoin.kt
@@ -11,10 +11,10 @@ import androidx.room.ForeignKey
     ]
 )
 class  DiaryHashtagJoin{
-    val diaryId: Int
+    val diaryId: Long
     val hashtagId: String
 
-    constructor(diaryId : Int , hashtagId : String){
+    constructor(diaryId : Long , hashtagId : String){
         this.diaryId = diaryId
         this.hashtagId = hashtagId
     }
